Added linkedin verification code support
Removed headless server suport and cleaned gui detection function

Added .DS_Store gitignore

Added back xvbf support to run headless driver, not bringing back headless auth

Fixed vault extra parameter error on platform.login()

Refactor

Refactor more

Added comment to specify virtual display import for headless setups, removed unused imports

Added openssl@1.1 dependency check for Darwin systems

Removed .DS_Store file from ./scripts

Platform refactor, LinkedIn verification, OpenSSL fix, headed support (#17)

* Fixed bug where ScraperEngine.build() created a default headless scraper w/ headed_support = None

* Fixed NoHeaded error, fixed infinite wait for LinkedIn verification, added new errors

* testing

* Refactored helper modulue, created new threader modulue to handle thread logging, fixed incorrect headed_support flag

* Refactored platform modulue. Platform subclasses reside in their own directory. Added universal selenium timeout constant

* Fixed incorrect openssl validation + changed gum install sequence

testing

Refactored helper modulue, created new threader modulue to handle thread logging, fixed incorrect headed_support flag

Added welcome print message + added script to `start.sh` (#18)

Welcome

Fixed bug where ScraperEngine.build() created a default headless scraper w/ headed_support = None

testing

Refactored helper modulue, created new threader modulue to handle thread logging, fixed incorrect headed_support flag

Refactored platform modulue. Platform subclasses reside in their own directory. Added universal selenium timeout constant

Fixed incorrect openssl validation + changed gum install sequence
