+++ scripts/utils/helpers.py
@@ -1,16 +1,9 @@
 import constants.constants as constants
-import multiprocessing
-import os
 import pathlib
-import traceback
-import logging
+import os
 import re
 import subprocess
 import time
-
-from . import logger
-from multiprocessing.pool import ThreadPool as Pool
-
 
 def has_gui() -> bool:
     """
@@ -74,7 +67,7 @@ def check_xvfb() -> bool:
                 ["sudo", "apt-get", "install", "xvfb"],
                 stderr=subprocess.DEVNULL,
             )
-            # Install firefox dependency 
+            # Install firefox dependency
             subprocess.check_call(
                 ["sudo", "apt-get", "install", "firefox"],
                 stderr=subprocess.DEVNULL,
@@ -93,7 +86,8 @@ def check_xvfb() -> bool:
             print("You are running on a server, addtional dependencies are required")
             input("Installing xserver-xephyr tigervnc-standalone-server x11-utils and gnumeric, press enter to continue...")
             subprocess.check_call(
-                ["sudo", "apt-get", "install", "xserver-xephyr", "tigervnc-standalone-server", "x11-utils", "gnumeric"],
+                ["sudo", "apt-get", "install", "xserver-xephyr",
+                    "tigervnc-standalone-server", "x11-utils", "gnumeric"],
                 stderr=subprocess.DEVNULL,
             )
             print("Installing additional python dependencies pyvirtualdisplay pillow and EasyProcess")
@@ -107,7 +101,7 @@ def check_xvfb() -> bool:
                 stderr=subprocess.DEVNULL,
             )
     return True
-        
+
 
 def get_root_from_url(url: str) -> str:
     """
@@ -118,43 +112,3 @@ def get_root_from_url(url: str) -> str:
     """
     regex = r"^(?:https?:\/\/)?(?:[^@\/\n]+@)?(?:www\.)?([^:\/?\n]+)"
     return re.findall(regex, url)[0]
-
-
-class LogExceptions(object):
-
-    """
-    Error logger for multiprocessing apply_async functions
-    """
-
-    def __init__(self, callable):
-        self.__callable = callable
-
-    def __call__(self, *args, **kwargs):
-        try:
-            result = self.__callable(*args, **kwargs)
-        except Exception as e:
-            # Here we add some debugging help. If multiprocessing's
-            # debugging is on, it will arrange to log the traceback
-            logger.LoggerBuilder.build(
-                log_level=logging.ERROR).error(traceback.format_exc()
-            )
-            LogExceptions.error(e)
-            # Re-raise the original exception so the Pool worker can
-            # clean up
-            raise
-
-        # It was fine, give a normal answer
-        return result
-
-    @staticmethod
-    def error(msg, *args):
-        return multiprocessing.get_logger().error(msg, *args)
-
-
-class LoggingPool(Pool):
-    """
-    Wrapper pool around logging exceptions
-    """
-
-    def apply_async(self, func, args=(), kwds={}, callback=None):
-        return Pool.apply_async(self, LogExceptions(func), args, kwds, callback)
